tests/conftest: Exit pytest if strategy transition fails
In case a transition in the strategy fails (and can not be recovered in
the strategy), we have to assume that this run is not recoverable.

With this change we log the exception and exit using `pytest.exit()`.
(`pytest.exit()` prints the exit reason in the last line of output what
comes in handy to identify the reason.)

This solution is not optimal: When exiting this way, a `junitxml` (or
similar) will contain a smaller number of tests (since some did not
run). This usually creates noise in the testing statistics.
A better solution would be to skip the missing tests with a useful
reason. But currently I am not aware of such solution.

Signed-off-by: Chris Fiege <chris@tinyhost.de>