ansible: deploy netplan network config
Use netplan as default way to configure the network for the
labgrid-controller. Stop if netplan is not available.
Netplan is available on major linux systems and doesn't require this to
deal with different ways of configure network on different Linux
distributions.

Add tasks to deploy netplan network config for netplan and a handler to
apply config with 'netplan apply'.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>